Add endpoint to create a new character (#80)
> **Notice:**
> **Endpoint to update a character level is probably broken right now
and will be fixed in a follow-up PR!**


Add an endpoint to create a new character (WIP! Does NOT work yet!):
- POST integration with API Gateway and step function
- Add Lambda function
- Add types and definitions for character sheet (e.g. list with combat
skills, etc.) --> necessary for character creation
- Add character factory to create a new, empty character sheet with
default values
- Update character and history schemas to make the types more precise
and usable for the character creation
- Add max string lengths for general character information
- Use max string length in different endpoints to avoid malformed
requests (e.g. add special ability)
- Add rules for character creation (e.g. start skills, effect of
hobby/profession, etc.)

Misc:
- **Add library with shared types/schemas of API for frontend and
backend code (WIP!)**
- Update update-level endpoint to use the new central, shared library.
It does work locally, but is probably broken in the cloud deployment atm
- Add functions to identify and log zod errors across different zod
versions and instances (when different packages throw zod errors)
- Make zod schemas more precise by explicitly defining an integer
instead of just a number for different endpoints
- Remove obsolete history record types in enum and reorder remaining
ones. Some things like hobby/profession, advantages, etc. are part of
the character creation and can't be changed independently for now ->
update set history record and revert history record endpoints (including
tests) + step function/state machine integrations
- Remove npm run command for zod schema creation (obsolete) -> the
schemas will be defined in the first place and build the single source
of truth
- Centrally log errors from HttpError
- fix: Update build scripts in backend/ to make them executable on MacOS
and Linux machines
- Minor changes as added TODO comments

Open TODO (will be solved in a follow-up PR):
- add tests
- add tests for add history record

+++ backend/src/lambdas/update-attribute/index.ts
@@ -23,13 +23,13 @@ export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayPr
 };
 
 const initialNewSchema = z.object({
-  initialValue: z.number(),
-  newValue: z.number(),
+  initialValue: z.number().int(),
+  newValue: z.number().int(),
 });
 
 const initialIncreasedSchema = z.object({
-  initialValue: z.number(),
-  increasedPoints: z.number(),
+  initialValue: z.number().int(),
+  increasedPoints: z.number().int(),
 });
 
 const bodySchema = z
